Breakup empty-attributes into definition and utils
