+++ src/options/empty-attributes/index.ts
@@ -1,6 +1,5 @@
 import { ChoiceSupportOption, PathArraySupportOption } from 'prettier';
-import { AttributeToken } from 'pug-lexer';
-import { CATEGORY_PUG } from '.';
+import { CATEGORY_PUG } from '..';
 
 export const PUG_EMPTY_ATTRIBUTES_OPTION: ChoiceSupportOption<PugEmptyAttributes> = {
 	since: '1.10.0',
@@ -36,41 +35,3 @@ export const PUG_EMPTY_ATTRIBUTES_FORCE_QUOTES_OPTION: PathArraySupportOption = 
 
 export type PugEmptyAttributes = 'as-is' | 'none' | 'all';
 export type PugEmptyAttributesForceQuotes = string[];
-
-const EMPTY_VALUES: [boolean, string, string] = [true, '""', "''"];
-
-export function formatEmptyAttribute(
-	token: AttributeToken,
-	pugEmptyAttributes: PugEmptyAttributes,
-	pugEmptyAttributesForceQuotes: PugEmptyAttributesForceQuotes
-): void {
-	const { val, name } = token;
-
-	const forceQuotesPatterns: RegExp[] = pugEmptyAttributesForceQuotes.map((pattern) => new RegExp(pattern));
-	const isForced: boolean = forceQuotesPatterns.some((pattern) => pattern.test(name));
-	if (isForced) {
-		if (token.val === true) {
-			token.val = '""';
-		}
-		return;
-	}
-
-	if (pugEmptyAttributes === 'as-is' || !EMPTY_VALUES.includes(val)) {
-		return;
-	}
-
-	switch (pugEmptyAttributes) {
-		case 'all': {
-			if (token.val === true) {
-				token.val = '""';
-			}
-			break;
-		}
-		case 'none': {
-			if (token.val === '""' || token.val === "''") {
-				token.val = true;
-			}
-			break;
-		}
-	}
-}
